packages/shared_utils/pyproject.toml
[build-system]
build-backend = "hatchling.build"

requires = [ "hatchling" ]

[project]
name = "shared-utils"
version = "0.1.0"
description = "Shared utilities for GrantFlow"
readme = "README.md"
license = { text = "UNLICENSED" }
authors = [ { name = "Na'aman Hirschfeld", email = "admin@grantflow.ai" } ]
requires-python = "==3.13.*"
classifiers = [ "Programming Language :: Python :: 3 :: Only", "Programming Language :: Python :: 3.13" ]

dependencies = [
  "anyio>=4.9",
  "msgspec>=0.19",
  "structlog>=25.2",
  "tenacity>=9",
]

optional-dependencies.ai = [
  "anthropic>=0.54",
  "google-genai>=0.1",
  "prompt-template>=1.2",
]
optional-dependencies.chunking = [
  "semantic-text-splitter>=0.27",
]
optional-dependencies.embeddings = [
  "sentence-transformers[onnx]>=4.1",
]
optional-dependencies.extraction = [
  "kreuzberg>=3.1.7",
]
optional-dependencies.gcs = [
  "google-cloud-storage>=2.19",
]
optional-dependencies.nlp = [
  "en-core-web-sm",
  "rank-bm25>=0.2.2",
  "spacy>=3.8.7",
  "torch>=2.7.1",
]
optional-dependencies.pubsub = [
  "google-cloud-pubsub>=2.30",
]

optional-dependencies.server = [
  "litestar[opentelemetry]>=2.16",
  "opentelemetry-exporter-gcp-monitoring>=1.6",
  "opentelemetry-exporter-gcp-trace>=1.6",
  "opentelemetry-instrumentation-httpx>=0.41b0",
  "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
  "uvicorn[standard]>=0.34.3",
]

[dependency-groups]
dev = [
  "db",
]

[tool.hatch.build.targets.wheel]
packages = [ "src" ]

[tool.hatch.build.targets.sdist]
packages = [ "src" ]

[tool.ruff]
lint.ignore = [ "PLC0415" ]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
